arrumei erros nos metodos e atributos

+++ Funcionario.java
@@ -13,7 +13,7 @@ public class Funcionario{
         return this.id;
     }
 
-    public int getSalario(){
+    public double getSalario(){
         return this.salario;
     }
